# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "FaultScenario": {
        "description": "A failure the gateway can inject (the Integration Lab); each has its own params.",
        "enum": [
          "auth_expired",
          "scope_denied",
          "rate_limited",
          "unsupported_parameter",
          "provider_5xx",
          "malformed_json",
          "truncated_stream",
          "slow_response",
          "timeout"
        ],
        "title": "FaultScenario",
        "type": "string"
      },
      "FaultScenarioCountOut": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "scenario": {
            "$ref": "#/components/schemas/FaultScenario"
          }
        },
        "required": [
          "scenario",
          "count"
        ],
        "title": "FaultScenarioCountOut",
        "type": "object"
      },
...
      "GatewayCacheOut": {
        "properties": {
          "hit_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hit Rate"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "misses": {
            "title": "Misses",
            "type": "integer"
          }
        },
        "required": [
          "hits",
          "misses",
          "hit_rate"
        ],
        "title": "GatewayCacheOut",
        "type": "object"
      },
...
      "GatewayKeyUsageOut": {
        "properties": {
          "cost_usd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "environment": {
            "title": "Environment",
            "type": "string"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "key_id": {
            "format": "uuid",
            "title": "Key Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          }
        },
        "required": [
          "key_id",
          "name",
          "environment",
          "requests",
          "errors",
          "cost_usd"
        ],
        "title": "GatewayKeyUsageOut",
        "type": "object"
      },
      "GatewayOverviewOut": {
        "description": "Gateway traffic of a window of at most 7 days.\n\n`errors` includes calls failed on purpose by Lab faults; `faults` says how many of the\nrequests carried one. `fallbacks` and `retries` are event counts, not request counts: the\nswitches to another target, and the repeated tries on one target, summed over all requests.\nA latency is `null` when no request measured it.",
        "properties": {
          "by_key": {
            "items": {
              "$ref": "#/components/schemas/GatewayKeyUsageOut"
            },
            "title": "By Key",
            "type": "array"
          },
          "by_target": {
            "items": {
              "$ref": "#/components/schemas/GatewayTargetUsageOut"
            },
            "title": "By Target",
            "type": "array"
          },
          "cache": {
            "$ref": "#/components/schemas/GatewayCacheOut"
          },
          "error_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Rate"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "fallbacks": {
            "title": "Fallbacks",
            "type": "integer"
          },
          "faults": {
            "items": {
              "$ref": "#/components/schemas/FaultScenarioCountOut"
            },
            "title": "Faults",
            "type": "array"
          },
          "p95_overhead_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Overhead Ms"
          },
          "p95_ttft_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ttft Ms"
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          },
          "retries": {
            "title": "Retries",
            "type": "integer"
          }
        },
        "required": [
          "requests",
          "errors",
          "error_rate",
          "cache",
          "fallbacks",
          "retries",
          "p95_overhead_ms",
          "p95_ttft_ms",
          "by_target",
          "by_key",
          "faults"
        ],
        "title": "GatewayOverviewOut",
        "type": "object"
      },
      "GatewayTargetUsageOut": {
        "properties": {
          "credential_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Id"
          },
          "credential_name": {
            "title": "Credential Name",
            "type": "string"
          },
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "p95_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ms"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "requests": {
            "title": "Requests",
            "type": "integer"
          }
        },
        "required": [
          "credential_id",
          "credential_name",
          "provider",
          "requests",
          "errors",
          "p95_ms"
        ],
        "title": "GatewayTargetUsageOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderKind": {
        "enum": [
          "openai",
          "anthropic",
          "openai_compatible"
        ],
        "title": "ProviderKind",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Spanlight API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/gateway/cache/purge": {
      "post": {
        "description": "Delete every cached response of the project, whatever key stored it.\n\nIdempotent: purging an empty cache succeeds. The audit event records how many entries went.",
        "operationId": "purge_cache_api_v1_projects__project_id__gateway_cache_purge_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge the gateway response cache",
        "tags": [
          "gateway"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/gateway/overview": {
      "get": {
        "description": "Requests, errors, cache, fallbacks, latency and usage per target and key.\n\nCounts every call that came through a gateway key, Lab faults included (`faults` says how\nmany). At most 7 days; longer is `422`.",
        "operationId": "overview_api_v1_projects__project_id__gateway_overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Environment"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayOverviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gateway traffic overview",
        "tags": [
          "gateway"
        ]
      }
    },
...
    "/gw/v1/chat/completions": {
      "post": {
        "description": "OpenAI Chat Completions through the key's route; streams when `stream` is true.",
        "operationId": "chat_completions_gw_v1_chat_completions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Chat Completions",
        "tags": [
          "gateway"
        ]
      }
    },
    "/gw/v1/messages": {
      "post": {
        "description": "Anthropic Messages through the key's route; streams when `stream` is true.",
        "operationId": "messages_gw_v1_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Messages",
        "tags": [
          "gateway"
        ]
      }
    },
...
    "/gw/v1/responses": {
      "post": {
        "description": "OpenAI Responses through the key's route; streams when `stream` is true.",
        "operationId": "responses_gw_v1_responses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Responses",
        "tags": [
          "gateway"
        ]
      }
    },
...
  }
}